"save/realtime+컴포넌트분리+roomname"
