+++ src/pages/Chat.tsx
@@ -6,39 +6,10 @@ import { useUserStore } from "../zustand/useUserStore";
 import { styled } from "styled-components";
 import { toast } from "react-toastify";
 
-interface Room {
-  room_id: string;
-  created_at: string;
-  roomname: string | null;
-}
-
 const Chat = () => {
   const { user } = useUserStore();
   const userId = user.userId; // users테이블의 userId를 user_id컬럼에 삽입, rpc에도 삽입
   const navigate = useNavigate();
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const form = e.currentTarget;
-    const { message } = Object.fromEntries(new FormData(form));
-
-    if (typeof message === "string" && message.trim().length !== 0) {
-      form.reset();
-      const { data, error } = await supabase
-        .from("messages")
-        .insert({ content: message, user_id: userId })
-        .select();
-      if (error) toast.error(error.message);
-
-      // rooms 테이블에 real-time 설정 안해주면 이 코드 적용해야 실시간반영됨
-      // const { data, error } = await supabase
-      //   .from("rooms")
-      //   .select("*")
-      //   .order("created_at", { ascending: false })
-      //   .limit(1)
-      //   .single();
-    }
-  };
 
   const handleCreateRoom = async () => {
     const { data, error } = await supabase.rpc("create_room", {
@@ -67,16 +38,6 @@ const Chat = () => {
             New room
           </button>
         </h1>
-        {/* 본문 */}
-        <Messages />
-        {/* 창 */}
-        <form onSubmit={handleSubmit} className="w-full bg-gray-100 p-1">
-          <input
-            type="text"
-            name="message"
-            className="w-full border-none bg-neutral-500"
-          />
-        </form>
       </div>
     </div>
   );
